requirements.txt
amqp==2.5.2
attrs==19.3.0
azure-core==1.5.0
azure-storage-blob==12.3.1
billiard==3.6.3.0
Brotli==1.0.7
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
more-itertools==8.3.0
msrest==0.6.14
mypy==0.770
mypy-extensions==0.4.3
numpy==1.18.4
oauthlib==3.1.0
packaging==20.4
pandas==1.0.3
plotly==4.8.1
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.2
pytest-mock==3.1.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
